FoxBook-py/pyproject.toml
[tool.poetry]
name = "foxbook"
version = "0.1.0"
description = ""
authors = ["JuliManhupli <juliamangup@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
